+++ src/decideid_assets/package.json
@@ -10,8 +10,9 @@
     "Canister"
   ],
   "scripts": {
-    "build": "node ../../scripts/build/gen_files_by_env.cjs && webpack",
-    "start": "export DEV_ENV=${DEV_ENV:-qa} &&  node ../../scripts/build/gen_files_by_env.cjs && webpack serve --mode development --config webpack.config.cjs"
+    "build": "npm run dynamically_gen_files && webpack",
+    "dynamically_gen_files": "node ../../scripts/build/gen_files_by_env.cjs",
+    "start": "export DEV_ENV=${DEV_ENV:-qa} &&  npm run dynamically_gen_files && webpack serve --mode development --config webpack.config.cjs"
   },
   "devDependencies": {
     "@dfinity/agent": "^0.19.3",
@@ -77,26 +78,19 @@
     "date-fns": "^2.25.0",
     "face-api.js": "^0.22.2",
     "final-form": "^4.20.4",
-    "htmlparser2": "^9.0.0",
     "ic-stoic-identity": "^2.0.0",
     "mic-recorder-to-mp3": "^2.2.2",
     "moment": "^2.29.4",
+    "path-to-regexp": "^6.2.1",
     "postcss-preset-env": "^9.3.0",
     "react": "^18.2.0",
-    "react-accessible-accordion": "^4.0.0",
-    "react-bulma-components": "^4.0.7",
-    "react-burger-menu": "^3.0.6",
-    "react-datepicker": "^4.8.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^11.4.2",
-    "react-easy-crop": "^4.1.4",
-    "react-final-form": "^6.5.7",
     "react-router": "^5.2.1",
     "react-router-dom": "^5.3.0",
     "react-router-hash-link": "^2.4.3",
     "react-scripts": "5.0.1",
     "react-webcam": "^6.0.0",
-    "semantic-ui-css": "^2.4.1",
     "styled-components": "^5.3.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^4.9.5",
