+++ src/decideid_assets/src/components/external/Home.tsx
@@ -4,6 +4,7 @@ import Landing from "./landing/Landing";
 export default function Home() {
   return (
     <div className="main" >
+      Landing Page
       <Landing />
     </div>
   )
