+++ src/decideid_assets/src/components/routers/External.tsx
@@ -1,16 +1,13 @@
 import * as React from "react";
 import { Route, Switch } from "react-router-dom";
-import Home from "./Home";
-import NewProfile from "../auth/new_profile/NewProfile";
-import Logout from "../auth/Logout";
+import Home from "../external/Home";
+
 
 export default function External() {
   return (
     <>
       <Switch>
         <Route exact path="/" component={Home} />
-        <Route path="/logout" component={Logout} />
-        <Route path="/signup" component={NewProfile} />
       </Switch>
     </>
   );
