+++ src/decideid_assets/src/index.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
-import { AuthProvider } from "./contexts/auth";
 import './globals.css';
 
 /**
@@ -11,9 +10,7 @@ window.global = window;
 
 ReactDOM.render(
   <React.StrictMode>
-    <AuthProvider>
       <App />
-    </AuthProvider>
   </React.StrictMode>,
   document.getElementById("app")
 );
